docker network config and further gql refactor
