+++ config/app_settings.py
@@ -2,18 +2,17 @@
 import os
 
 
-SECRET_KEY = os.getenv('SECRET_KEY', None)
+SECRET_KEY = os.getenv('SECRET_KEY')
 
-SERVER_NAME = os.getenv(
-    'SERVER_NAME', 'localhost:{0}'.format(os.getenv('PORT', '8000')))
+SERVER_NAME = os.getenv('SERVER_NAME')
 
 # MongoDB configuration
 MONGODB_SETTINGS = {
     'db': os.getenv('MONGODB_DB', 'sqrl'),
     'host': os.getenv('MONGODB_HOST', 'mongodb'),
     'port': int(os.getenv('MONGODB_PORT', 27017)),
-    'username': os.getenv('MONGODB_USERNAME', None),
-    'password': os.getenv('MONGODB_PASSWORD', None),
+    'username': os.getenv('MONGODB_USERNAME'),
+    'password': os.getenv('MONGODB_PASSWORD'),
 }
 
 GATOR_CLIENT_URL = os.getenv('GATOR_CLIENT_URL')
