+++ sqrl/graphql/schemas/timetable.py
@@ -5,7 +5,7 @@ import graphene
 
 from gator.models.timetable import Course
 
-from sqrl.graphql.objects.timetable import UserTimetableObject
+from sqrl.graphql.objects import UserTimetableObject
 from sqrl.models import UserTimetable
